+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# REPLACE_GLOBALLY_WITH_NEXT_VERSION
+- PPI-246 - Added option for advanced logging
+
 # 3.0.3
 - PPI-20 - Fix Webhooks when payment status is already set identically
 - PPI-235 - Fix Webhooks not correctly registered for separate Sales Channel credentials